tipo de expediente: ver y editar
se reutilizo el modal editar de objeto de gasto
