+++ frontend/package.json
@@ -23,7 +23,6 @@
     "react-scripts": "3.4.0",
     "react-spring": "^8.0.27",
     "react-select": "^3.1.0",
-    "react-spring": "^8.0.27",
     "styled-components": "^5.1.1"
   },
   "scripts": {
