+++ frontend/src/services/TiposDeExpedientes.js
@@ -5,12 +5,20 @@ class TiposDeExpedientes {
     return http.get('/tipos_de_expedientes/?format=json');
   }
 
+  getById(id) {
+    return http.get(`/tipos_de_expedientes/${id}?format=json`);
+  }
+
   getDetails(id) {
-    return http.get('/tipos_de_expedientes_detalles/?format=json&tipo_de_expediente_id=' + id);
+    return http.get(`/tipos_de_expedientes_detalles/?format=json&tipo_de_expediente_id=${id}`);
   }
 
   create(tipoExpediente) {
     return http.post('/tipos_de_expedientes/', tipoExpediente);
+  }
+
+  update(id, tipoExpediente) {
+    return http.put(`/tipos_de_expedientes/${id}`, tipoExpediente);
   }
 
   createDetail(detail) {
@@ -18,7 +26,7 @@ class TiposDeExpedientes {
   }
 
   delete(id) {
-    return http.delete('tipos_de_expedientes', id);
+    return http.delete(`tipos_de_expedientes/${id}`);
   }
 }
